// src/types/formTypes.ts
export interface FormData {
  email: string;
  currentFinancialInstitution: string;
  lookingFor: string;
  religiousOrganization: string;
  shariaCompliant: boolean;
  currentEmployer: string;
  studentOrAlumni: string;
  currentOrFormerMilitary: string;
  militaryBranch: string;
  environmentalInitiatives: boolean;
  diversityEquityInclusion: boolean;
  religion: string;
}